Update postsrsd.c
Some OSes like e.g. the BSDs have the connection inherit O_NONBLOCK.
The result does not prevent postsrsd from working at all, but causes a syslogged error for roughly every other connection from cleanup.
Removing the flag is a NOP if it's not set, so this change doesn't require an OS-dependent ifdef.