Move milter state creation from connect callback to envfrom
If IPv6 support is not enabled in libmilter, the connect callback is not called
when an IPv6 client connects. Work around the issue by moving state
initialization from the connect callback to the envelope-from callback.

Note that the state is now per-message, not per-connection, so it must be
cleaned up in the end-of-message callback, in the abort callback, and in any
other callback that finalizes message handling (e.g., tempfail).